tests/wpt/web-platform-tests/html/canvas/element/transformations/2d.transformation.perspective.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.transformation.perspective</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.transformation.perspective</h1>
<p class="desc">perspective() results in the correct transformation matrix</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("perspective() results in the correct transformation matrix");
_addTest(function(canvas, ctx) {

const length = 100;
ctx.perspective(length);
const domMatrix = new DOMMatrix();
domMatrix.m34 = -1/length;
_assertMatricesApproxEqual(domMatrix, ctx.getTransform());
ctx.rotateAxis(1, 2, 3, 4);
domMatrix.rotateAxisAngleSelf(1, 2, 3, rad2deg(4));
_assertMatricesApproxEqual(domMatrix, ctx.getTransform());


});
</script>
